more reorganizing and no more lodash usage
